backend/pyproject.toml
[tool.poetry]
name = "twit-backend"
version = "0.1.0"
description = ""
authors = ["Seah Chin Ying <scy.6char@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.55.1"}
databases = {extras = ["postgresql"], version = "^0.3.2"}
python-dotenv = "^0.13.0"
psycopg2-binary = "^2.8.5"
"marrow.mailer" = "^4.0.3"
redis = "^3.5.2"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"
flake8 = "^3.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
